backend/requirements.txt
Flask>=2.0
Flask-Cors>=3.0
gunicorn>=20.0
google-cloud-firestore>=2.5
google-generativeai>=0.3
requests>=2.25
beautifulsoup4>=4.9
